Moved IgnoredChannels, ServerWarningInfo, UlongRoleId, UlongUserId, UserActivityContainer to Models.SubModules
