+++ Railgun/Commands/Filters/AntiUrl.cs
@@ -9,6 +9,7 @@ using Railgun.Core.Commands.Attributes;
 using Railgun.Core.Utilities;
 using TreeDiagram;
 using TreeDiagram.Models.Server.Filter;
+using TreeDiagram.Models.SubModels;
 
 namespace Railgun.Commands.Filters
 {
