+++ Railgun/Commands/Inactivity/Ignore.cs
@@ -5,6 +5,7 @@ using Finite.Commands;
 using Railgun.Core.Commands;
 using TreeDiagram;
 using TreeDiagram.Models.Server;
+using TreeDiagram.Models.SubModels;
 
 namespace Railgun.Commands.Inactivity
 {
