+++ Railgun/Commands/Inactivity/InactivityMonitor.cs
@@ -11,6 +11,7 @@ using System.Text;
 using System.Threading.Tasks;
 using TreeDiagram;
 using TreeDiagram.Models.Server.Inactivity;
+using TreeDiagram.Models.SubModels;
 
 namespace Railgun.Commands.Inactivity
 {
