+++ Railgun/Commands/Music/MusicRoleLock.cs
@@ -6,6 +6,7 @@ using Railgun.Core.Commands;
 using Railgun.Core.Commands.Attributes;
 using TreeDiagram;
 using TreeDiagram.Models.Server;
+using TreeDiagram.Models.SubModels;
 
 namespace Railgun.Commands.Music
 {
