+++ Railgun/Core/Events.cs
@@ -16,6 +16,7 @@ using TreeDiagram;
 using TreeDiagram.Enums;
 using TreeDiagram.Models.Server;
 using TreeDiagram.Models.Server.Inactivity;
+using TreeDiagram.Models.SubModels;
 
 namespace Railgun.Core
 {
