+++ Railgun/Core/Managers/InactivityManager.cs
@@ -8,6 +8,7 @@ using System.Timers;
 using Railgun.Core.Containers;
 using TreeDiagram;
 using TreeDiagram.Models.Server.Inactivity;
+using TreeDiagram.Models.SubModels;
 using TreeDiagram.Models.TreeTimer;
 
 namespace Railgun.Core.Managers
