+++ TreeDiagram/Models/Server/Filter/FilterBase.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using TreeDiagram.Models.SubModels;
 
 namespace TreeDiagram.Models.Server.Filter
 {
