+++ TreeDiagram/Models/Server/Warning/ServerWarning.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
+using TreeDiagram.Models.SubModels;
 
 namespace TreeDiagram.Models.Server.Warning
 {
